refactor: Product 필드 값 객체로 변경에 따른 ProductEntity 생성 및 Dao, Repository 구분

+++ src/main/java/cart/dto/ProductDto.java
@@ -1,27 +1,29 @@
 package cart.dto;
 
+import java.math.BigDecimal;
+
 public class ProductDto {
 
     private final Long productId;
     private final String name;
     private final String image;
-    private final Long price;
+    private final BigDecimal price;
 
-    public ProductDto(String name, String image, long price) {
+    public ProductDto(String name, String image, BigDecimal price) {
         this.productId = null;
         this.name = name;
         this.image = image;
         this.price = price;
     }
 
-    public ProductDto(Long productId, String name, String image, long price) {
+    public ProductDto(Long productId, String name, String image, BigDecimal price) {
         this.productId = productId;
         this.name = name;
         this.image = image;
         this.price = price;
     }
 
-    public long getProductId() {
+    public Long getProductId() {
         return productId;
     }
 
@@ -33,7 +35,7 @@ public class ProductDto {
         return image;
     }
 
-    public long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
